Display type + id when type is not supported
