+++ src/pages/index.vue
@@ -1,3 +1,11 @@
+<script lang="ts">
+import Vue from 'vue';
+
+export default Vue.extend({
+  name: 'PageIndex',
+});
+</script>
+
 <template>
   <q-page class="flex flex-center">
     Home
@@ -6,11 +14,3 @@
 
 <style>
 </style>
-
-<script lang="ts">
-import Vue from 'vue';
-
-export default Vue.extend({
-  name: 'PageIndex',
-});
-</script>
